refactor(ipc): rename sync-handlers → sync-core-handlers (Phase 2.4)
After splitting OAuth, device/OTP, and attachment concerns into their
own handler files, what remains in the former god-module is pure
sync-engine control (status/trigger/history/queue/pause/resume/
emergency-wipe/etc.) plus startup integrity and the cross-cutting
clearInMemoryAuthState composer. Rename the file to reflect that
narrower role and delete the back-compat import path — memry is
pre-production, so the old path has no consumers outside this branch.

- apps/desktop/src/main/ipc/sync-handlers.ts → sync-core-handlers.ts
- apps/desktop/src/main/ipc/sync-handlers.test.ts → sync-core-handlers.test.ts
- Update imports: ipc/index.ts, ipc/index.test.ts (vi.mock path),
  sync/session-teardown.ts
- Regenerate generated-ipc-invoke-map.ts + preload/generated-rpc.ts
