+++ apps/desktop/src/main/ipc/sync-core-handlers.ts
