+++ apps/desktop/src/main/sync/session-teardown.ts
@@ -10,7 +10,7 @@ import { resetTokenManagerState } from './token-manager'
 import { getValidAccessToken } from './token-manager'
 import { clearPendingSession, clearPendingLinkCompletion } from './linking-service'
 import { getCrdtProvider, resetCrdtProvider } from './crdt-provider'
-import { clearInMemoryAuthState } from '../ipc/sync-handlers'
+import { clearInMemoryAuthState } from '../ipc/sync-core-handlers'
 import { getDatabase, isDatabaseInitialized } from '../database/client'
 import { store } from '../store'
 import { createLogger } from '../lib/logger'
